fix: some tests
