+++ src/components/InternalDropdown/helpers/__test__/isChildRegistered.spec.js
