+++ src/components/InternalDropdown/helpers/__test__/isOptionVisible.spec.js
