+++ src/components/InternalDropdown/helpers/scrollTo.js
@@ -1,0 +1,3 @@
+export default function scrollTo(ref, offset) {
+    ref.current.scrollTo(0, offset);
+}
